worked on the main page and added the contact page
Things are coming together!!!! Working on this site is so much fun. Made a whole animation for the logo. Seems to work fine in chrome, seems to not work the way I want it to in firefox but its alright :) The skill bars on the about page is working mighty fine! Stoked on that one. Shout out to this guy:https://www.youtube.com/watch?v=EoN_4fv3T7A&ab_channel=CascadingStyle